Program for Inverse of an 'n' Ordered Matrix
I tried to create a program which returns the inverse of the Entered Matrix. However, C cannot hold fractional expressions in variables. I tried changing the format into float but still there are many inaccuracies. Creating this program in JavaScript would be way easier because we have an 'eval' function which can evaluate a string if it contains math expressions. Looking forward to bring an updated version of this program in JavaScript.
This program can apply a row interchange operation with perfection.